Major diff.
It seems that this patch brings us to an acceptable version.

Some TODOs:
We build a new resolver for every domain each time. This is stupid -- we could loop after we have built all our domain resolvers.
Make sure we don't leak memory, some of ldns* frees are omitted.
We can't free the char pointer that builds the hostname (we're missing something ;-))